proper unit test for hypemachine / track metadata extraction
